Rescue console broken pipe
If the 2d application isn't running then we'll get an error trying to
write to it when using the console, we rescue that and direct the user
to the documentation.

While testing this I noticed that the program you're running doesn't
output to stdout, so I imagine users will hit this issue fairly
frequently unless they can see the output of their programs. I had a
typo in my require and the only output I saw was this message as the
program had already exited due to the exception and It didn't show me
why.