Rough refactor text rendering code
A very rough refactoring of the code used to render text, this should
hopefully provide an idea of where we can go. There's lots of
improvements to be made and this doesn't seem to render just yet, but it
should be good enough to figure out if this is the direction we want to
go.