Refactor font->texture loading and re-use fonts across text objects
Now multiple text objects can use the same font, so we don't have to use
up a file handle for every single one, this should help with performance
and resources in general.

Also update the API around Text and the way we interact with fonts and
textures to be more idiomatic ruby. Initially I wasn't using a font
instance, rather just storing our C font object inside of text, but by
having a font instance it cleans up the interface from text-> font
nicely.

Also the responsibility of creating a texture from a font was previously
on the font class, now that responsibility is in the Texture class.
Adding `Texture.load_text` tidies up this API as well!