Add new texture drawing method to openGL2 and openGLES
I have manually tested openGL2 and it works fine, I can't test openGL ES
as it's not currently working.

In this patch I've refactored the `Vertices` class in ruby as previously
it was tied to the openGL 3 code's vertices array structure, with
everything bundled together. Now it presents co-ordinates, texture
co-ordinates and color information all separately for the individual
rendering functions to use to make the structures they need.

Vertices no longer does anything with color information, that is now
directly passed by the `Text` class.