Refactor Texture drawing and re-add text resizing
The texture class now holds onto an SDL surface object (rgb pixel data
in RAM) when it's created, then later we create a texture based on that
(rgb pixel data but stored in GPU memory rather than RAM).

Now that we get a surface early on, we can use it to extract the
width/height of the font which fixes our tests which were broken from
before. It also allows us to easily re-size text by creating a new
texture.