Add ability to text to be resized
When we change the size value of our text, we need to create a new 'font'
object as well as creating a new texture which holds the pixel data of the text
drawn out.

I have moved these two things "creating the font" and "creating the texture"
into new private methods on the text class for reuse.