Remove some unused code
This condition would check if we failed to create an image and present a
helpful error message. We don't use the same C code for loading images
anymore and so it may be better to handle this inside of the Texture
class, as Texture will get a blank array instead of useful arguments.