Tileset rendering refactor (#211)
* Update tileset to use `Image` and `Texture` for rendering

Also add the ability to flip and rotate tiles

Lastly extract the logic to check if an image exists to the `Image`
class for re-use.

* Remove C tileset code

We don't need this any more as we're using `Image` and `Texture` to
render out tiles now :)

* Update tileset flipping and rotation

Use keyword arguments so the user can easily specify either flip or rotation.
Add these to the tileset rendering test to ensure it doesn't break :)

* Test setting the window icon in the render test

I'd like to factor the code which sets the window icon, so let's test it
first :)

* Update window icon to use new rendering, remove R2D_Image

The window icon was the last place we were using the R2D_Image C struct. I've
updated this code to use the new image functions to directly get a surface and
removed the remaining unused image code.

* Remove now unused rendering methods

These methods were used by Sprite/Tileset/Text and Image, but now they
use the new function and directly provide vertices information. So these
can be removed.

* Rename new draw texture to be just 'texture' now

This texture rendering function now replaces the old one, so we can
remove the 'new' word.