Add scaling to tileset and fix spacing bug
Previously: to scale the size of tiles displayed in a tileset, you would need
to multiply the tile width/height, the image width/height as well as the
spacing and padding values by the scale. This works but means that you need to
know the dimensions of the tileset image, rather than having it be detected and
you have to multiply each value which can be confusing.

This patch still allows you to manually set those values if you wish but
introduces a 'scale' attribute which will do the work for you, meaning now you
can set a scale without having to specify the width/height of the tileset image
manually and multiply all attributes.

This patch also fixes a bug where the spacing was not being calculated
correctly, it was using `(tile_x + spacing) * tile_width` which is incorrect,
it is now: `tile_x * (spacing + tile_width)`