Add error handling around open windows (#215)
* Add error handling around open windows

Add two errors:
1) prevent users from calling 'show' multiple times as it currently
   doesn't function
2) prevent users directly drawing objects outside of the render block
   (this does not work as the window is not yet ready)

* Add @blacktm's suggestions for detecting multiple open windows

* Ensure all renderable classes do a window check

For some reason I had missed these classes earlier. Also remove one call
in Triangle to the window check, as we only need to perform it on direct
draw calls, not when the private render method is called (as the window
is always ready in that case).