Update ruby2d.rb
If you use a non-RubyInstaller version of Ruby (IE from the likes of MSYS2), the "RubyInstaller" class is not available.

This prevents Ruby2D from running your scripts, citing a missing `ruby2d.so` file.

The reason this was happening is because of the following line: `RubyInstaller::Runtime.add_dll_directory(File.expand_path(s2d_dll_path))`

The above line adds the `assets/mingw/bin` folder to an include list (so Ruby2D can use the DLL's it contains), but is dependent on `RubyInstaller`. The fix is to get rid of this dependency, which can be done using the `fiddle` gem. It seems that `fiddle` is a default Ruby gem, so everybody should have it installed.

The above code works for me. It's not been tested on other versions of Ruby.