Add the ability to get/remove and replace existing tiles
When the tileset feature was added, the idea was that you would clear
all of the tiles before painting new ones, this proves to be costly from
a performance standpoint, so instead it's easier if we can sub out a
single tile for a new one.

The tilesets are kept in an array which we need to scan to remove tiles,
I think at the moment this will likely be ok for performance, we can sub
out the array for a hash keyed by the x+y coordinates of the tile;
though retrieving the hash values for rendering might incur some memory
copy cost as well, so happy to change it if we feel necessary.